Dev to UAT merge

See merge request root/employeeportal_java!36

+++ src/main/java/com/apmosys/employeeportal/repository/CompOffMasterRepository.java
@@ -6,6 +6,6 @@ import com.apmosys.employeeportal.model.CompOffMaster;
 
 public interface CompOffMasterRepository extends JpaRepository<CompOffMaster, Short> {
 
-	CompOffMaster findByCompOffId(Integer reasonId);
+	CompOffMaster findByCompOffId(short reasonId);
 
 }
